refactored auction house for forestfire
