+++ Assets/Scripts/AuctionHouseExternalities.cs
@@ -1,0 +1,46 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using UnityEngine;
+using UnityEngine.Assertions;
+using AYellowpaper.SerializedCollections;
+using UnityEngine.Rendering;
+using Sirenix.OdinInspector;
+using ChartAndGraph;
+using EconSim;
+using Sirenix.Serialization;
+using Sirenix.OdinInspector.Editor.ValueResolvers;
+using UnityEngine.Serialization;
+
+public partial class AuctionHouse 
+{
+	bool forestFire = false;
+	[HideIf("forestFire")]
+	[Button(ButtonSizes.Large), GUIColor(0.4f, 0.8f,1)]
+	public void ForestFire()
+	{
+		//TODO rapid decline (*.6 every round) for 2-5 rounds, then regrow at 1.1 until reaches back to 1 multiplier
+		//do this in a new class
+		var wood = district.book["Wood"];
+		var weight = wood.productionMultiplier;
+		weight = .5f;
+
+		wood.ChangeProductionMultiplier(weight);
+		wood.productionChance = 0.5f;
+
+		forestFire = true;
+	}
+	[ShowIf("forestFire")]
+	[Button(ButtonSizes.Large), GUIColor(1, 0.4f, 0.4f)]
+	public void StopForestFire()
+	{
+		var wood = district.book["Wood"];
+		var weight = wood.productionMultiplier;
+		weight = 1f;
+		wood.ChangeProductionMultiplier(weight);
+		wood.productionChance = 1f;
+
+		forestFire = false;
+	}
+}
